Move kernel cmdline routines to util package
