--- pkg/config/read_test.go
@@ -1,22 +1,0 @@
-package config
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestParseCmdLine(t *testing.T) {
-	cmdline := `x y harvester.a.b=true "harvester.c=d" harvester.e harvester.f=1 harvester.f=2`
-	m, err := parseCmdLine(cmdline, kernelParamPrefix)
-	if err != nil {
-		t.Fatal(err)
-	}
-	want := map[string]interface{}{
-		"a": map[string]interface{}{"b": "true"},
-		"c": "d",
-		"e": "true",
-		"f": []string{"1", "2"},
-	}
-	assert.Equal(t, want, m)
-}
